# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (Admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        status:
          type: string
          enum: [active, archived]
        ownerClerkId:
          type: string
        taskCount:
          type: integer
        memberCount:
          type: integer
        completedCount:
          type: integer
        myRole:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - status
        - ownerClerkId
        - taskCount
        - memberCount
        - completedCount
        - myRole
        - createdAt
        - updatedAt
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/ProjectMember"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
          required:
            - members
            - tasks
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
      required:
        - name
        - color
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        status:
          type: string
          enum: [active, archived]
    ProjectMember:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - userId
        - clerkId
        - name
        - email
        - role
        - joinedAt
...
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        projectColor:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        assigneeId:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        assigneeAvatarUrl:
          type: string
          nullable: true
        assigneeClerkId:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - projectName
        - projectColor
        - title
        - status
        - priority
        - createdAt
        - updatedAt
...